+++ src/Machinery/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [Unreleased]
+### Added
+- Constraint `IDisposable` for `TState` type parameter of `DisposableStateMachine<…>` generics.
+
 ## [0.3.0] - 2019-10-16
 ### Added
 - `DisposableStateMachine<…>` classes owning their states.
@@ -22,6 +26,7 @@
 ### Added
 - Basic building blocks for generic state machines: `StateMachine<…>`, `IState<…>`, `IPolicy<…>`.
 
+[Unreleased]: https://github.com/qbit86/machinery/compare/machinery-0.3.0...HEAD
 [0.3.0]: https://github.com/qbit86/machinery/compare/machinery-0.2.0...machinery-0.3.0
 [0.2.0]: https://github.com/qbit86/machinery/compare/machinery-0.1.1...machinery-0.2.0
 [0.1.1]: https://github.com/qbit86/machinery/compare/machinery-0.1.0...machinery-0.1.1